+++ ccm.h
@@ -55,13 +55,13 @@
  * \param la  The number of additional authentication octets (may be zero).
  * \return FIXME
  */
-long int
+size_t
 dtls_ccm_encrypt_message(rijndael_ctx *ctx, size_t M, size_t L, 
 			 unsigned char nonce[DTLS_CCM_BLOCKSIZE], 
 			 unsigned char *msg, size_t lm, 
 			 const unsigned char *aad, size_t la);
 
-long int
+size_t
 dtls_ccm_decrypt_message(rijndael_ctx *ctx, size_t M, size_t L, 
 			 unsigned char nonce[DTLS_CCM_BLOCKSIZE], 
 			 unsigned char *msg, size_t lm,
