+++ crypto.c
@@ -180,7 +180,7 @@ dtls_ccm_encrypt(aes128_ccm_t *ccm_ctx, const unsigned char *src, size_t srclen,
 		 unsigned char *buf, 
 		 unsigned char *nounce,
 		 const unsigned char *aad, size_t la) {
-  long int len;
+  size_t len;
 
   assert(ccm_ctx);
 
@@ -197,7 +197,7 @@ dtls_ccm_decrypt(aes128_ccm_t *ccm_ctx, const unsigned char *src,
 		 size_t srclen, unsigned char *buf,
 		 unsigned char *nounce,
 		 const unsigned char *aad, size_t la) {
-  long int len;
+  size_t len;
 
   assert(ccm_ctx);
